Fix request entity issues
- One issue was that the signature validation on the server side failed because the request
body on the client was the String representation of a Java object and the request body on the
server was the serialized representation (e.g. json). This caused signatures on the client and
server to be different. This is now fixed in the client. The client now gets the request body
(if anything is there) and serializes it so that the serialized representation can be included
when generating the client-side signature. This will then enable the client and server signatures
to match.

- The client and server modules now support binary content in the request body by treating it as
as byte array instead of a String.

- Added new unit tests to cover both of these.