Ignore the host and port in the path
When calculating the signature in the client or server, the path is now included as
the portion of the URI after the port (e.g. /api/1/pizza?apiKey=someKey&foo=bar).
Previously, it included the parts before it as well, but this causes problems in
some cases and is not actually needed.